SEBankIDQRCode fullscreen support
to support accessibility requirements (https://developers.bankid.com/getting-started/qr-code#accessibility)
this commit adds support for clicking the QR code to show it in
fullscreen mode.

a few breaking changes are made, such as converting the canvas to
an image tag, for better rerender support when swapping
between fullscreen and not.